config/*:fix failed test caused by failpoint
