+++ config/config_util_test.go
@@ -23,7 +23,6 @@ import (
 
 	"github.com/BurntSushi/toml"
 	. "github.com/pingcap/check"
-	"github.com/pingcap/failpoint"
 )
 
 func (s *testConfigSuite) TestCloneConf(c *C) {
@@ -168,22 +167,4 @@ engines = ["tikv", "tiflash", "tidb"]
 	c.Assert(toJSONStr(flatMap["log.level"]), Equals, `"info"`)
 	c.Assert(toJSONStr(flatMap["log.format"]), Equals, `"text"`)
 	c.Assert(toJSONStr(flatMap["isolation-read.engines"]), Equals, `["tikv","tiflash","tidb"]`)
-}
-
-func (s *testConfigSuite) TestTxnScopeValue(c *C) {
-	failpoint.Enable("github.com/pingcap/tidb/config/injectTxnScope", `return("bj")`)
-	isGlobal, v := GetTxnScopeFromConfig()
-	c.Assert(isGlobal, IsFalse)
-	c.Assert(v, Equals, "bj")
-	failpoint.Disable("github.com/pingcap/tidb/config/injectTxnScope")
-	failpoint.Enable("github.com/pingcap/tidb/config/injectTxnScope", `return("")`)
-	isGlobal, v = GetTxnScopeFromConfig()
-	c.Assert(isGlobal, IsTrue)
-	c.Assert(v, Equals, "global")
-	failpoint.Disable("github.com/pingcap/tidb/config/injectTxnScope")
-	failpoint.Enable("github.com/pingcap/tidb/config/injectTxnScope", `return("global")`)
-	isGlobal, v = GetTxnScopeFromConfig()
-	c.Assert(isGlobal, IsFalse)
-	c.Assert(v, Equals, "global")
-	failpoint.Disable("github.com/pingcap/tidb/config/injectTxnScope")
 }
