+++ store/tikv/config/config_test.go
@@ -15,6 +15,7 @@ package config
 
 import (
 	. "github.com/pingcap/check"
+	"github.com/pingcap/failpoint"
 )
 
 var _ = SerialSuites(&testConfigSuite{})
@@ -30,4 +31,23 @@ func (s *testConfigSuite) TestParsePath(c *C) {
 	_, disableGC, err = ParsePath("tikv://node1:2379?disableGC=true")
 	c.Assert(err, IsNil)
 	c.Assert(disableGC, IsTrue)
+}
+
+func (s *testConfigSuite) TestTxnScopeValue(c *C) {
+
+	failpoint.Enable("github.com/pingcap/tidb/store/tikv/config/injectTxnScope", `return("bj")`)
+	isGlobal, v := GetTxnScopeFromConfig()
+	c.Assert(isGlobal, IsFalse)
+	c.Assert(v, Equals, "bj")
+	failpoint.Disable("github.com/pingcap/tidb/store/tikv/config/injectTxnScope")
+	failpoint.Enable("github.com/pingcap/tidb/store/tikv/config/injectTxnScope", `return("")`)
+	isGlobal, v = GetTxnScopeFromConfig()
+	c.Assert(isGlobal, IsTrue)
+	c.Assert(v, Equals, "global")
+	failpoint.Disable("github.com/pingcap/tidb/store/tikv/config/injectTxnScope")
+	failpoint.Enable("github.com/pingcap/tidb/store/tikv/config/injectTxnScope", `return("global")`)
+	isGlobal, v = GetTxnScopeFromConfig()
+	c.Assert(isGlobal, IsFalse)
+	c.Assert(v, Equals, "global")
+	failpoint.Disable("github.com/pingcap/tidb/store/tikv/config/injectTxnScope")
 }
